requirements.txt
Pillow==10.2.0
numpy>=1.21.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6 